♻️ Extract the fixture object into an abstract layer and move the test suite of clickup operations into another new test module.
